requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
eventlet==0.36.1
psycopg2-binary==2.9.9

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0

